+++ src/components/recipe/RecipePage.vue
@@ -85,7 +85,7 @@
                     <template v-if="$slots.instructions || instructions.length > 0">
                         <h2>{{ $t('recipes.instructions') }}</h2>
                         <slot name="instructions">
-                            <ol v-if="recipe && recipe.instructionSteps.length > 0" class="pl-0 list-none">
+                            <ol v-if="recipe && recipe.instructionStepUrls.length > 0" class="pl-0 list-none">
                                 <li
                                     v-for="(instructionStep, index) of instructions"
                                     :key="instructionStep.position"