+++ src/models/RecipesListItem.schema.ts
@@ -1,0 +1,10 @@
+import { FieldType } from 'soukai';
+import { SolidModel } from 'soukai-solid';
+
+export default SolidModel.schema({
+    recipeUrl: {
+        type: FieldType.Key,
+        rdfProperty: 'schema:item',
+        required: true,
+    },
+});