+++ src/services/CookbookService.ts
@@ -1,5 +1,6 @@
-import { PromisedValue, arr, isObject, tap, uuid } from '@noeldemartin/utils';
-import { SolidContainerModel } from 'soukai-solid';
+import { PromisedValue, arr, isObject, requireUrlParentDirectory, tap, uuid } from '@noeldemartin/utils';
+import { SolidContainerModel, SolidTypeRegistration } from 'soukai-solid';
+import { SolidDocumentPermission, findInstanceRegistrations } from '@noeldemartin/solid-utils';
 import type { FluentArray, Obj } from '@noeldemartin/utils';
 
 import Auth from '@/framework/core/facades/Auth';
@@ -10,8 +11,9 @@ import Service from '@/framework/core/Service';
 import { setRemoteCollection } from '@/framework/cloud/remote_helpers';
 import type { ComputedStateDefinitions, IService } from '@/framework/core/Service';
 
-import Recipe from '@/models/Recipe';
 import Network from '@/framework/core/facades/Network';
+import Recipe from '@/models/Recipe';
+import RecipesList from '@/models/RecipesList';
 
 declare module '@/framework/core/services/EventsService' {
 
@@ -26,6 +28,7 @@ interface State {
     localCookbookUrl: string;
     remoteCookbookUrl: string | null;
     cookbook: PromisedValue<SolidContainerModel>;
+    publicRecipesList: RecipesList | null;
     allRecipes: FluentArray<Recipe>;
 }
 
@@ -45,7 +48,7 @@ export default class CookbookService extends Service<State, ComputedState> {
         return this.cookbook.isResolved();
     }
 
-    public async createRemote(name: string, storageUrl: string): Promise<void> {
+    public async initializeRemote(name: string, storageUrl: string): Promise<void> {
         const user = Auth.requireUser();
         const engine = Auth.requireAuthenticator().engine;
         const cookbook = await SolidContainerModel.withEngine(engine, async () => tap(
@@ -58,14 +61,22 @@ export default class CookbookService extends Service<State, ComputedState> {
             },
         ));
 
-        await this.setRemoteCookbook(cookbook);
+        await this.initializeRemoteCookbook(cookbook);
         await this.reloadRecipes();
+    }
+
+    public getTmpRecipe(id: string): Recipe | null {
+        return this.tmpRecipes[id] ?? null;
+    }
+
+    public saveTmpRecipe(recipe: Recipe): string {
+        return tap(uuid(), id => this.tmpRecipes[id] = recipe);
     }
 
     public async deleteRecipe(recipe: Recipe): Promise<void> {
         if (this.isRemote) {
             await recipe.softDelete();
-            await (Network.online && Cloud.sync());
+            await (Network.online && Cloud.sync(recipe));
 
             return;
         }
@@ -77,12 +88,38 @@ export default class CookbookService extends Service<State, ComputedState> {
         await recipe.delete();
     }
 
-    public saveTmpRecipe(recipe: Recipe): string {
-        return tap(uuid(), id => this.tmpRecipes[id] = recipe);
-    }
+    public async updatePublicRecipeListing(recipe: Recipe, publish: boolean): Promise<void> {
+        const publicRecipesList = await this.loadPublicRecipesList();
 
-    public getTmpRecipe(id: string): Recipe | null {
-        return this.tmpRecipes[id] ?? null;
+        if (publish && !publicRecipesList.has(recipe)) {
+            publicRecipesList.relatedItems.attach({ recipeUrl: recipe.url });
+
+            recipe.listUrls = [...recipe.listUrls, publicRecipesList.url];
+        }
+
+        if (!publish && publicRecipesList.has(recipe)) {
+            const item = publicRecipesList.items?.find(item => item.recipeUrl === recipe.url);
+
+            item && publicRecipesList.relatedItems.detach(item);
+
+            recipe.listUrls = recipe.listUrls.filter(url => url !== publicRecipesList.url);
+        }
+
+        if (publicRecipesList.exists() || (publicRecipesList.items && publicRecipesList.items.length > 0)) {
+            await RecipesList.withEngine(
+                Auth.requireAuthenticator().engine,
+                () => publicRecipesList.save(),
+            );
+        }
+
+        if (publicRecipesList.exists() && publicRecipesList.wasRecentlyCreated() && !publicRecipesList.isRegistered) {
+            await this.registerPublicRecipesList(publicRecipesList);
+        }
+
+        if (recipe.isDirty()) {
+            await recipe.save();
+            await Cloud.sync(recipe);
+        }
     }
 
     protected async boot(): Promise<void> {
@@ -92,19 +129,23 @@ export default class CookbookService extends Service<State, ComputedState> {
         await this.loadRecipes();
         await this.enqueueFileUploads();
 
-        Events.on('login', async () => {
+        Events.on('login', async session => {
             // TODO this should be refactored, because technically the cookbook is already loaded (it's just the local
             // one, not the remote, that's why this works now). Instead we should migrate both the local cookbook and
             // the recipes to match the remote declarations.
             await this.loadCookbook();
+
+            RecipesList.setEngine(session.authenticator.engine);
+            SolidTypeRegistration.setEngine(session.authenticator.engine);
         });
 
         Events.on('logout', async () => {
             Recipe.collection = this.localCookbookUrl;
 
             await Promise.all(this.allRecipes.map(async recipe => {
-                if (recipe.imageUrl?.startsWith(this.localCookbookUrl))
+                if (recipe.imageUrl?.startsWith(this.localCookbookUrl)) {
                     await Files.delete(recipe.imageUrl);
+                }
 
                 await recipe.delete();
             }));
@@ -136,6 +177,7 @@ export default class CookbookService extends Service<State, ComputedState> {
             localCookbookUrl: Recipe.collection,
             remoteCookbookUrl: null,
             cookbook: new PromisedValue,
+            publicRecipesList: null,
             allRecipes: arr<Recipe>([]),
         };
     }
@@ -148,35 +190,7 @@ export default class CookbookService extends Service<State, ComputedState> {
         };
     }
 
-    private async loadRecipes(): Promise<void> {
-        const recipes = await Recipe.all();
-
-        this.allRecipes = arr(recipes);
-    }
-
-    protected async reloadRecipes(): Promise<void> {
-        await this.loadRecipes();
-    }
-
-    private async loadCookbook(): Promise<void> {
-        if (Auth.isLoggedIn()) {
-            const engine = Auth.authenticator.engine;
-            const cookbook = await SolidContainerModel.withEngine(engine, async () => tap(
-                await this.findCookbook(),
-                cookbook => cookbook && this.setRemoteCookbook(cookbook),
-            ));
-
-            cookbook && this.cookbook.resolve(cookbook);
-
-            return;
-        }
-
-        Recipe.collection = this.remoteCookbookUrl ?? this.localCookbookUrl;
-
-        this.cookbook.resolve(new SolidContainerModel({ url: Recipe.collection, name: 'cookbook' }));
-    }
-
-    private async setRemoteCookbook(cookbook: SolidContainerModel): Promise<void> {
+    private async initializeRemoteCookbook(cookbook: SolidContainerModel): Promise<void> {
         await this.migrateLocalRecipes(cookbook.url);
 
         setRemoteCollection(Recipe, cookbook.url);
@@ -187,13 +201,50 @@ export default class CookbookService extends Service<State, ComputedState> {
         this.cookbook.resolve(cookbook);
     }
 
+    private async loadRecipes(): Promise<void> {
+        const recipes = await Recipe.all();
+
+        this.allRecipes = arr(recipes);
+    }
+
+    private async loadCookbook(): Promise<void> {
+        if (Auth.isLoggedIn()) {
+            const engine = Auth.authenticator.engine;
+            const cookbook = await SolidContainerModel.withEngine(engine, async () => tap(
+                await this.findCookbook(),
+                cookbook => cookbook && this.initializeRemoteCookbook(cookbook),
+            ));
+
+            cookbook && this.cookbook.resolve(cookbook);
+
+            return;
+        }
+
+        Recipe.collection = this.remoteCookbookUrl ?? this.localCookbookUrl;
+        RecipesList.collection = this.remoteCookbookUrl ?? RecipesList.collection;
+
+        this.cookbook.resolve(new SolidContainerModel({ url: Recipe.collection, name: 'cookbook' }));
+    }
+
+    private async loadPublicRecipesList(): Promise<RecipesList> {
+        this.publicRecipesList = this.publicRecipesList
+            ?? await this.findPublicRecipesList()
+            ?? this.newPublicRecipesList();
+
+        return this.publicRecipesList;
+    }
+
+    private async reloadRecipes(): Promise<void> {
+        await this.loadRecipes();
+    }
+
     private async findCookbook(refreshStaleProfile: boolean = true): Promise<SolidContainerModel | null> {
         const user = Auth.requireUser();
 
         try {
-            return user.privateTypeIndexUrl
-                ? await SolidContainerModel.fromTypeIndex(user.privateTypeIndexUrl, Recipe)
-                : null;
+            return await this.findCookbookInTypeIndex(user.privateTypeIndexUrl)
+                ?? await this.findCookbookInTypeIndex(user.publicTypeIndexUrl)
+                ?? null;
         } catch (error) {
             // TODO check that error was NetworkRequestError
             if (!refreshStaleProfile)
@@ -204,6 +255,56 @@ export default class CookbookService extends Service<State, ComputedState> {
 
             return this.findCookbook(false);
         }
+    }
+
+    private async findCookbookInTypeIndex(typeIndexUrl?: string): Promise<SolidContainerModel | null> {
+        if (!typeIndexUrl) {
+            return null;
+        }
+
+        return SolidContainerModel.fromTypeIndex(typeIndexUrl, Recipe);
+    }
+
+    private async findPublicRecipesList(): Promise<RecipesList | null> {
+        if (!Auth.isLoggedIn()) {
+            return null;
+        }
+
+        const user = Auth.requireUser();
+
+        if (user.publicTypeIndexUrl) {
+            const authenticator = Auth.requireAuthenticator();
+            const engine = authenticator.engine;
+            const urls = await findInstanceRegistrations(
+                user.publicTypeIndexUrl,
+                RecipesList.rdfsClasses,
+                authenticator.requireAuthenticatedFetch(),
+            );
+
+            for (const url of urls) {
+                const list = await RecipesList.withEngine(engine, () => RecipesList.find(url));
+
+                if (!list?.isPublic) {
+                    continue;
+                }
+
+                return list;
+            }
+        }
+
+        return null;
+    }
+
+    private newPublicRecipesList(): RecipesList {
+        const user = Auth.requireUser();
+        const remoteCookbookUrl = this.remoteCookbookUrl
+                ?? fail('Cannot create a recipes list without a remote cookbook');
+
+        return new RecipesList({
+            url: `${remoteCookbookUrl}public#it`,
+            name: user.name ? `${user.name}'s Public Recipes` : 'Public Recipes',
+            creator: user.webId,
+        });
     }
 
     private async migrateLocalRecipes(remoteCollection: string): Promise<void> {
@@ -268,6 +369,33 @@ export default class CookbookService extends Service<State, ComputedState> {
 
             Cloud.enqueueFileUpload(url);
         }
+    }
+
+    private async registerPublicRecipesList(publicRecipesList: RecipesList): Promise<void> {
+        const user = Auth.requireUser();
+        const engine = Auth.requireAuthenticator().engine;
+        const typeIndexUrl = user.publicTypeIndexUrl ?? await Auth.createPublicTypeIndex();
+
+        await Promise.all(RecipesList.rdfsClasses.map(async rdfsClass => {
+            const typeRegistration = new SolidTypeRegistration({
+                forClass: rdfsClass,
+                instance: publicRecipesList.url,
+            });
+
+            typeRegistration.mintUrl(typeIndexUrl, true, uuid());
+
+            await SolidTypeRegistration.withEngine(
+                engine,
+                () => typeRegistration.save(requireUrlParentDirectory(typeIndexUrl)),
+            );
+        }));
+
+        await RecipesList.withEngine(
+            engine,
+            () => publicRecipesList.updatePublicPermissions([SolidDocumentPermission.Read]),
+        );
+
+        publicRecipesList.isRegistered = true;
     }
 
 }